FileIO and Logging used in MEDYS-Java Projects, can be usefull in other Projects.
Some Methodnames are in Englisch some in German as those classes were
written for internal purpose.

There are many FileIO-Projects or Libraries, so this one should not only
be the first choice!

See : Apache IO (Commons), Java NIO, and many others not mentioned here!

Same goes for the Logging!

Cheers,

emisan (hek)